# Add 1PC optimization to delete/update.