@tamird 2016-02-15
LGTM

---

Reviewed 4 of 4 files at r1.
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[sql/bench_test.go, line 243 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4384#-KAafI-5BdWOQ-ZF9U01:-KAafI-5BdWOQ-ZF9U02:1855034637) ([raw file](https://github.com/cockroachdb/cockroach/blob/a96d4172b0b3f662b341a228194dc94a1e97beee/sql/bench_test.go#L243)):*
Couldn't you parse everything and batch whatever's between a BEGIN and a COMMIT?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4384#-:-KAaf_zgEobP5Ohmbr9q:676834233)*
<!-- Sent from Reviewable.io -->