@petermattis 2016-02-15
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---

*[sql/bench_test.go, line 243 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4384#-KAafI-5BdWOQ-ZF9U01:-KAaiAfwt7RY9D3Cc854:-157284160) ([raw file](https://github.com/cockroachdb/cockroach/blob/a96d4172b0b3f662b341a228194dc94a1e97beee/sql/bench_test.go#L243)):*
a) Getting the error handling for the correct would be tricky and b) We actually need something more than "batching". We'd want to buffer up all of the writes while still performing the reads needed by each update. 

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4384)*
<!-- Sent from Reviewable.io -->