@petermattis 2016-02-15
@tschottdorf Any thoughts about that `Delete100` slow down? The diff between the two perf numbers amounts to the following code:

```
		pErr = p.txn.Run(b)
		if pErr == nil {
			pErr = p.txn.Commit()
		}
```

vs

```
		pErr = p.txn.CommitInBatch(b)
```

Surprisingly, the separate `Run` and `Commit` is the faster case for `Delete100`. A cpu profile shows that using `CommitInBatch` we spend more time in RocksDB calls, though I have yet to determine why.