@petermattis 2016-02-15
I'm going to merge this and file an issue to track down what caused the `Delete100` slow down.